Not Able to Save Province or Change Working Location
There was a small glith in the process for adding a new employee. The user could not change the location for any employee while adding them. They also could not change their province. This was due to the post back on the page load. Corrected here and on the Customer as it was the same issue.