Bug Fix #004
Users are able to complete a sale on a return receipt.  By starting a return then closing the browser the user could then locate the return invoice in the current open sales and start a sale process. this has now been fixed. The only problem now to correct is that those invoices are still in the current table. Will continue to work on a way to have those removed. This update will currently fix this issue for now.