Trade In Total
When doing a return on a trade in the amount was not reversing on the trade in total for that day. So on a day when 1 trade in is taken the returned back to customer the total should equal 0 but it wasn't. corrected this error and made sure that when doing a return the application was recording the employee who did it and not the admin logged in.